feat(bridge-exec): generate bridge proof via ZK program
